refactor: rewrite svg into file for importing and managing them
