+++ skirunmap/src/images/FilterSVG/index.tsx
