+++ skirunmap/src/images/ShowOrHideArrowSVG/index.tsx
@@ -1,0 +1,27 @@
+const ShowOrHideArrowSVG = ({ isShown }: { isShown: boolean }) => {
+  return isShown ? (
+    <svg
+      xmlns='http://www.w3.org/2000/svg'
+      fill='none'
+      viewBox='0 0 24 24'
+      strokeWidth='1.5'
+      stroke='currentColor'
+      className='h-6 w-6'
+    >
+      <path strokeLinecap='round' strokeLinejoin='round' d='M4.5 15.75l7.5-7.5 7.5 7.5'></path>
+    </svg>
+  ) : (
+    <svg
+      xmlns='http://www.w3.org/2000/svg'
+      fill='none'
+      viewBox='0 0 24 24'
+      strokeWidth='1.5'
+      stroke='currentColor'
+      className='h-6 w-6'
+    >
+      <path strokeLinecap='round' strokeLinejoin='round' d='M19 9l-7 7-7-7'></path>
+    </svg>
+  )
+}
+
+export default ShowOrHideArrowSVG
