+++ skirunmap/src/images/UploadUserIconSVG/index.tsx
@@ -1,0 +1,20 @@
+const UploadFileSVG = () => {
+  return (
+    <svg
+      xmlns='http://www.w3.org/2000/svg'
+      fill='none'
+      viewBox='0 0 24 24'
+      strokeWidth='1.5'
+      stroke='white'
+      className='absolute inset-1/2 z-10 h-8 w-8 -translate-x-1/2 -translate-y-1/2'
+    >
+      <path
+        strokeLinecap='round'
+        strokeLinejoin='round'
+        d='M3 16.5v2.25A2.25 2.25 0 005.25 21h13.5A2.25 2.25 0 0021 18.75V16.5m-13.5-9L12 3m0 0l4.5 4.5M12 3v13.5'
+      />
+    </svg>
+  )
+}
+
+export default UploadFileSVG
