+++ skirunmap/src/pages/RouteEdit/index.tsx
@@ -14,6 +14,8 @@ import { v4 as uuidv4 } from 'uuid'
 import { db, storage } from '../../auth/Firebase'
 import Map from '../../components/Map'
 import Modal from '../../components/Modal'
+import ShowOrHideArrowSVG from '../../images/ShowOrHideArrowSVG'
+import UploadGpxFileSVG from '../../images/UploadGpxFileSVG'
 import { MarkerWithSpotId, useMapStore } from '../../store/useMap'
 import { Route, Spot, useRouteStore, useSpotStore } from '../../store/useRoute'
 import { useUserStore } from '../../store/useUser'
@@ -562,21 +564,7 @@ const EditRoute: React.FC = () => {
                     isDragOver && 'text-zinc-300'
                   }`}
                 >
-                  <svg
-                    xmlns='http://www.w3.org/2000/svg'
-                    fill='none'
-                    viewBox='0 0 24 24'
-                    strokeWidth='1.5'
-                    stroke='currentColor'
-                    className='h-6 w-6'
-                  >
-                    <path
-                      strokeLinecap='round'
-                      strokeLinejoin='round'
-                      d='M19.5 14.25v-2.625a3.375 3.375 0 00-3.375-3.375h-1.5A1.125 1.125 0 0113.5 7.125v-1.5a3.375 3.375 0 00-3.375-3.375H8.25m6.75 12l-3-3m0 0l-3 3m3-3v6m-1.5-15H5.625c-.621 0-1.125.504-1.125 1.125v17.25c0 .621.504 1.125 1.125 1.125h12.75c.621 0 1.125-.504 1.125-1.125V11.25a9 9 0 00-9-9z'
-                    />
-                  </svg>
-
+                  <UploadGpxFileSVG />
                   <label htmlFor='gpxFile' className='cursor-pointer rounded-l-2xl pl-2'>
                     Drag GPX file or&nbsp;
                   </label>
@@ -630,33 +618,7 @@ const EditRoute: React.FC = () => {
               onClick={() => toggleRouteVisibility()}
             >
               <p className='text-3xl font-bold'>Route</p>
-              {routeVisibility ? (
-                <svg
-                  xmlns='http://www.w3.org/2000/svg'
-                  fill='none'
-                  viewBox='0 0 24 24'
-                  strokeWidth='1.5'
-                  stroke='rgb(59,130,246)'
-                  className='h-6 w-6'
-                >
-                  <path
-                    strokeLinecap='round'
-                    strokeLinejoin='round'
-                    d='M4.5 15.75l7.5-7.5 7.5 7.5'
-                  ></path>
-                </svg>
-              ) : (
-                <svg
-                  xmlns='http://www.w3.org/2000/svg'
-                  fill='none'
-                  viewBox='0 0 24 24'
-                  strokeWidth='1.5'
-                  stroke='rgb(59,130,246)'
-                  className='h-6 w-6'
-                >
-                  <path strokeLinecap='round' strokeLinejoin='round' d='M19 9l-7 7-7-7'></path>
-                </svg>
-              )}
+              <ShowOrHideArrowSVG isShown={routeVisibility} />
               <div className='w-full border border-blue-400' />
             </div>
 
@@ -788,33 +750,7 @@ const EditRoute: React.FC = () => {
               onClick={() => toggleSpotsVisibility()}
             >
               <p className='text-3xl font-bold'>Spots</p>
-              {spotVisibility ? (
-                <svg
-                  xmlns='http://www.w3.org/2000/svg'
-                  fill='none'
-                  viewBox='0 0 24 24'
-                  strokeWidth='1.5'
-                  stroke='rgb(59,130,246)'
-                  className='h-6 w-6'
-                >
-                  <path
-                    strokeLinecap='round'
-                    strokeLinejoin='round'
-                    d='M4.5 15.75l7.5-7.5 7.5 7.5'
-                  ></path>
-                </svg>
-              ) : (
-                <svg
-                  xmlns='http://www.w3.org/2000/svg'
-                  fill='none'
-                  viewBox='0 0 24 24'
-                  strokeWidth='1.5'
-                  stroke='rgb(59,130,246)'
-                  className='h-6 w-6'
-                >
-                  <path strokeLinecap='round' strokeLinejoin='round' d='M19 9l-7 7-7-7'></path>
-                </svg>
-              )}
+              <ShowOrHideArrowSVG isShown={spotVisibility} />
               <div className='w-full border border-blue-400' />
             </div>
